mail-poller/requirements.txt
async-timeout==4.0.3
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
kafka-python==2.0.2
redis==5.0.2
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
